+++ poultry/producers.py
@@ -21,7 +21,6 @@ def consume_stream(target, input_dir=None):
         with consumers.closing(*targets):
 
             for line in lines:
-
                 if not line.strip():
                     continue
 
@@ -48,14 +47,13 @@ def readline_dir(input_dir, extract_retweets=False, mark_extracted=False):
         try:
             tweet = Tweet(l)
         except TweetValueError:
-            pass
+            continue
         else:
 
             # TODO: this duplicates consumers.extract_retweets.
             retweeted_status = tweet.parsed.get('retweeted_status', None)
             if extract_retweets and retweeted_status:
                     yield Tweet(retweeted_status)
-
             yield tweet