+++ poultry/tweet.py
@@ -25,7 +25,7 @@ class Tweet(object):
             except ValueError:
                 raise TweetValueError("The passed json can't be parsed.")
             else:
-                if isinstance(tweet, dict) and 'text' in tweet:
+                if isinstance(tweet, dict) and ('text' in tweet or 'full_text' in tweet):
                     self.raw = raw_json
                     self.parsed = tweet
                 else:
@@ -34,7 +34,7 @@ class Tweet(object):
     @property
     def text(self):
         """The unprocessed text of the tweet."""
-        return self.parsed['text']
+        return self.parsed['full_text'] if 'full_text' in self.parsed else self.parsed['text']
 
     @property
     def hashtags(self):
@@ -216,11 +216,15 @@ class Tweet(object):
         tokens = self.text_without_entities.split()
         tokens = (t.lower() for t in tokens)
 
-        return filter(lambda s: len(s) >= min_token_len,
-                      [u''.join(c for c in t if unicodedata.category(c)[0] in allowed_categories)
-                       for t in tokens
-                       ]
-                      )
+        return list(
+            filter(
+                lambda s: len(s) >= min_token_len,
+                [
+                    u''.join(c for c in t if unicodedata.category(c)[0] in allowed_categories)
+                    for t in tokens
+                ]
+            )
+        )
 
     @property
     def is_spam(self):
@@ -229,7 +233,7 @@ class Tweet(object):
 
         .. todo: parameters instead of magic numbers.
         '''
-        entity_violations = (2 < e for e in [self.hashtags, self.urls, self.user_mentions])
+        entity_violations = (len(e) > 2 for e in [self.hashtags, self.urls, self.user_mentions])
         lenght_violantion = len(self.tokens) < 5
 
         return any(entity_violations) or lenght_violantion